Remove "enabled" field from UrlEncodedFormCondition
There is no need for this. If a request has a body that contains
something other than form parameters, the "URLEncodedUtils.parse" method
will not attempt to parse it and will return an empty list. It will only
parse the body if the request has a "application/x-www-form-urlencoded"
content type.